% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretrends-plot.R
\name{slope_for_power}
\alias{slope_for_power}
\title{Linear trend for which pre-trends tests have a given power level}
\usage{
slope_for_power(
  sigma,
  targetPower = 0.5,
  tVec,
  referencePeriod = 0,
  prePeriodIndices = which(tVec < referencePeriod)
)
}
\arguments{
\item{sigma.}{The covariance matrix for the event-plot coefficients.}

\item{targetPower.}{The desired power for the pre-test, which reject if any pre-treatment coefficient is significant at the 5% level}

\item{tVec.}{The vector of time periods corresponding with the coefficients}

\item{referencePeriod.}{(Optional) The omitted pre-treatment reference period normalized to 0. Default is t=0}

\item{prePeriodIndices.}{(Optional) The indices of the event-plot corresponding with pre-treatment periods. Default is which(tVec<referencePeriod)}
}
\value{
slope. The slope of the trend for which targetPower is achieved.
}
\description{
This function computes the linear violation of parallel trends for which convention pre-trends tests have a pre-specified power level
}
